eks_distributed_scraper/slave/requirements.txt
requests
beautifulsoup4
pyzmq
